# openapi.yaml
openapi: 3.0.3
info:
  title: Vetios AI Operations Platform API
  version: '2026-04-05'
  description: Core contract for inference, evaluation, governance, telemetry, webhooks, datasets, and simulations.
servers:
  - url: https://app.vetios.ai
security:
  - bearerAuth: []
    tenantScope: []
paths:
...
  /api/evaluation:
    get:
      summary: List evaluations and summary metrics
      parameters:
        - { in: query, name: tenant_id, schema: { type: string } }
        - { in: query, name: model_version, schema: { type: string } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { description: Evaluation list, content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope' } } } }
    post:
      summary: Trigger evaluation for an outcome or inference
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EvaluationCreateRequest' }
      responses:
        '200': { description: Evaluation created, content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope' } } } }
        '429': { description: Tenant rate limit exceeded, content: { application/json: { schema: { $ref: '#/components/schemas/RateLimitExceeded' } } } }
...
  /api/governance/policy:
    get:
      summary: List governance policies
      parameters:
        - { in: query, name: tenant_id, schema: { type: string } }
      responses:
        '200': { description: Policy list, content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope' } } } }
    post:
      summary: Create a draft governance policy
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GovernancePolicyCreateRequest' }
      responses:
        '201': { description: Policy created, content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope' } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    tenantScope: { type: apiKey, in: header, name: X-Tenant-Scope }
  schemas:
    ApiEnvelope:
      type: object
      properties:
        data:
          nullable: true
          oneOf:
            - { type: object, additionalProperties: true }
            - { type: array, items: { type: object, additionalProperties: true } }
            - { type: string }
            - { type: integer }
            - { type: number }
            - { type: boolean }
        meta: { $ref: '#/components/schemas/ApiMeta' }
        error: { $ref: '#/components/schemas/ApiError', nullable: true }
    ApiMeta:
      type: object
      properties:
        tenant_id: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        version: { type: string }
        request_id: { type: string }
        next_cursor: { type: string, nullable: true }
        limit: { type: integer }
    ApiError:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
    RateLimitExceeded:
      type: object
      properties:
        error: { type: string, enum: [rate_limit_exceeded] }
        tenant_id: { type: string }
        limit: { type: integer }
        window_seconds: { type: integer }
        retry_after_seconds: { type: integer }
...
    EvaluationCreateRequest:
      type: object
      required: [model_name, model_version]
      properties:
        outcome_id: { type: string, format: uuid }
        inference_event_id: { type: string, format: uuid }
        model_name: { type: string }
        model_version: { type: string }
        predicted_confidence: { type: number }
        trigger_type: { type: string, enum: [inference, outcome, simulation] }
...
    GovernancePolicyCreateRequest:
      type: object
      properties:
        tenant_id: { type: string, nullable: true }
        name: { type: string }
        rules: { type: object, additionalProperties: true }
        metadata: { type: object, additionalProperties: true }
...